+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# v1.0.0
+## 04/07/2017
+1. [](#improved)
+    * enable auto fill for keyword metadata
+1. [](#bugfix)
+    * Fixed several issue before public release (see commit list for details)
+
 # v0.1.0
 ## 04/07/2017